+++ TableProTests/PluginTestSources/PostgreSQLCatalogPresence.swift
@@ -1,0 +1,1 @@
+../../Plugins/PostgreSQLDriverPlugin/PostgreSQLCatalogPresence.swift